requirements.txt
django==3.1.5
djangorestframework==3.12.2
requests==2.25.1
beautifulsoup4==4.9.3
psycopg2==2.8.4